+++ src/NewLoanScreen.java
@@ -67,7 +67,7 @@ public class NewLoanScreen extends JDialog {
             JOptionPane.showMessageDialog(contentPane, "You need to provide some collateral.");
             return;
         }
-        if (!atm.requestLoan(new BigDecimal((int) amountSpinner.getValue()), currencyCombo.getSelectedItem().toString(), collateralText.getText())) {
+        if (!atm.requestLoan(new BigDecimal((Double) amountSpinner.getValue()), currencyCombo.getSelectedItem().toString(), collateralText.getText())) {
             JOptionPane.showMessageDialog(contentPane, "Failed to request loan");
             return;
         }
